+++ packages/common/src/core/schemas/internal/getEditorSchemas.ts
@@ -29,6 +29,7 @@ import { getCardEditorSchemas } from "./getCardEditorSchemas";
 import { SurveyEditorType } from "../../../surveys/_internal/SurveySchema";
 import { EventEditorType } from "../../../events/_internal/EventSchemaCreate";
 import { UserEditorType } from "../../../users/_internal/UserSchema";
+import { addDynamicSlugValidation } from "./addDynamicSlugValidation";
 
 /**
  * get the editor schema and uiSchema defined for an editor (either an entity or a card).
@@ -458,6 +459,9 @@ export async function getEditorSchemas(
 
   // filter out properties not used in the UI schema
   schema = filterSchemaToUiSchema(schema, uiSchema);
+
+  // if schema includes slug, add conditional validation
+  schema = addDynamicSlugValidation(schema, options);
 
   return Promise.resolve({ schema, uiSchema, defaults });
 }