+++ packages/common/src/permissions/types/Permission.ts
@@ -17,9 +17,6 @@ import { UserPermissions } from "../../users/_internal/UserBusinessRules";
  * It's critical that the arrays defined in the modules use `as const`
  * otherwise Permission devolves into just a string type
  */
-// this is a temporary mechanism for gating workspaces or parts of workspaces
-// to be used until we release workspaces or it can be replaced with our new access control (permission/feature/capability) system
-const TempPermissions = ["temp:workspace:released"];
 
 const SystemPermissions = [
   "hub:feature:privacy",
@@ -28,9 +25,6 @@ const SystemPermissions = [
   "hub:card:follow",
   "hub:feature:workspace:user",
   "hub:feature:workspace:org",
-  // DEPRECATED: This permission has been replaced by hub:feature:workspace:org,
-  // remove this at the next breaking version
-  "hub:feature:workspace:umbrella",
   "hub:feature:keyboardshortcuts",
   "hub:feature:newentityview",
   "hub:feature:history",
@@ -56,7 +50,6 @@ const validPermissions = [
   ...GroupPermissions,
   ...PagePermissions,
   ...PlatformPermissions,
-  ...TempPermissions,
   ...DiscussionPermissions,
   ...InitiativeTemplatePermissions,
   ...TemplatePermissions,
@@ -77,7 +70,6 @@ export type Permission =
   | (typeof GroupPermissions)[number]
   | (typeof PagePermissions)[number]
   | (typeof PlatformPermissions)[number]
-  | (typeof TempPermissions)[number]
   | (typeof DiscussionPermissions)[number]
   | (typeof InitiativeTemplatePermissions)[number]
   | (typeof TemplatePermissions)[number]