+++ packages/common/src/sites/feed-configuration.ts
@@ -1,8 +1,11 @@
 import { IModel } from "../types";
+import { getMajorVersion } from "./feeds/_internal/getMajorVersion";
 
 type FeedFormat = "dcat-us" | "dcat-ap" | "rss";
 
 /**
+ * DEPRECATED: Use `getFeedTemplate()` instead
+ * this can be deleted once we remove the feed editor from the content-library (Ember addon)
  * Returns feed configuration from a site model
  *
  * @param {IModel} site - site model
@@ -30,6 +33,7 @@ export function getFeedConfiguration(
 }
 
 /**
+ * DEPRECATED: This will be removed in the next breaking version. Use `setFeedTemplate()` instead;
  * Returns feed configuration from a site model
  *
  * @param {IModel} site - site model
@@ -122,8 +126,4 @@ function setRssConfig(
   }
 
   throw new Error("Unsupported RSS Version");
-}
-
-function getMajorVersion(version: string) {
-  return version.split(".")[0];
 }