+++ packages/common/src/utils/index.ts
@@ -18,7 +18,6 @@ export * from "./increment-string";
 export * from "./logger";
 export * from "./is-update-group";
 export * from "./revertable-tasks";
-export * from "./sessionLocalStorage";
 export * from "./dasherize";
 export * from "./titleize";
 export * from "./memoize";