+++ packages/common/test/initiatives/HubInitiative.test.ts
@@ -1,6 +1,6 @@
 import * as PortalModule from "@esri/arcgis-rest-portal";
 import { IHubInitiative, IMetric, IResolvedMetric, getProp } from "../../src";
-import { Catalog } from "../../src/search";
+import { Catalog } from "../../src";
 import { ArcGISContextManager } from "../../src/ArcGISContextManager";
 import { HubInitiative } from "../../src/initiatives/HubInitiative";
 import { MOCK_AUTH } from "../mocks/mock-auth";
@@ -533,40 +533,6 @@ describe("HubInitiative Class:", () => {
         // since thumbnailCache is protected we can't really test that it's set
         // other than via code-coverage
         expect(getProp(result, "_thumbnail")).not.toBeDefined();
-      });
-      it("handles extent from location", async () => {
-        const chk = HubInitiative.fromJson(
-          {
-            id: "bc3",
-            name: "Test Entity",
-            thumbnailUrl: "https://myserver.com/thumbnail.png",
-            extent: [
-              [-1, -1],
-              [1, 1],
-            ],
-          },
-          authdCtxMgr.context
-        );
-        // spy on the instance .save method and retrn void
-        const saveSpy = spyOn(chk, "save").and.returnValue(Promise.resolve());
-        // make changes to the editor
-        const editor = await chk.toEditor();
-        editor.name = "new name";
-        editor.location = {
-          extent: [
-            [-2, -2],
-            [2, 2],
-          ],
-          type: "custom",
-        };
-        // call fromEditor
-        const result = await chk.fromEditor(editor);
-        // expect the save method to have been called
-        expect(saveSpy).toHaveBeenCalledTimes(1);
-        expect(result.extent).toEqual([
-          [-2, -2],
-          [2, 2],
-        ]);
       });
       it("handles setting featured image", async () => {
         const chk = HubInitiative.fromJson(